refactor(app): make on device display deck configurator updates live
Instead of the previous method of staging a batch of changes to deck configuration before commiting
them with the confirm button, to avoid synchronization issues, make the ODD deck configurator behave
like the desktop version. all updates immediately permeate to the robot's source of truth
